You can return anything in the import_u64_async (very broken)
It seems you can return any type you like in the import_u64_async implementation in the runtime. It still compiles fine. It just hangs at runtime:

test test::async_import_and_export has been running for over 60 seconds

Any test that run before this run fine. It also seems to work fine for u32 (and possibly u16 and u8), even returning the correct value at runtime as u64 in the wasm module, but still hangs for other primitives or for structs.

+++ examples/example-rust-wasmer-runtime/src/spec/mod.rs
@@ -143,8 +143,8 @@ async fn import_string_async() -> String {
     "Imported string".to_owned()
 }
 
-async fn import_u64_async() -> u64 {
-    64
+async fn import_u64_async() -> f32 {
+    64.0
 }
 
 fn log(msg: String) {
